fix: appease copilot review

+++ cpp/include/monoprop/MonomialPropagator.h
@@ -242,14 +242,14 @@ public:
                      const VecD &gen_coeffs,
                      std::optional<VecZ> gate_indices = std::nullopt,
                      std::optional<VecD> parameters = std::nullopt,
-                     std::optional<int> only_rotate_len_k = std::nullopt) -> void;
+                     std::optional<size_t> only_rotate_len_k = std::nullopt) -> void;
 
     /// Evolve and contract immediately, without storing a graph.
     auto propagate(const std::vector<VecZ> &majoranas,
                    const VecZ &parameter_mapping,
                    const VecD &gen_coeffs,
                    const VecD &parameters,
-                   std::optional<int> only_rotate_len_k = std::nullopt) -> void;
+                   std::optional<size_t> only_rotate_len_k = std::nullopt) -> void;
 
     auto expectation_value(const VecD &parameters) -> double;
 
@@ -410,7 +410,7 @@ private:
                                   const VecZ &parameter_mapping,
                                   const VecD &gen_coeffs,
                                   const VecZ &gate_indices,
-                                  std::optional<int> only_rotate_len_k) -> void;
+                                  std::optional<size_t> only_rotate_len_k) -> void;
 
     // Returns {build_angle, apply_angle}; apply is the build angle, negated in Schrödinger.
     auto gate_angle_(const VecD &mapped_params, size_t i, size_t majoranas_size) const -> std::pair<double, double> {
@@ -425,21 +425,21 @@ private:
                                         const VecZ &gate_indices,
                                         const VecD &parameters,
                                         const VecD &operator_coeffs,
-                                        std::optional<int> only_rotate_len_k) -> void;
+                                        std::optional<size_t> only_rotate_len_k) -> void;
 
     auto evolve_mode_contract_immediately_(const std::vector<VecZ> &majoranas,
                                            const VecZ &parameter_mapping,
                                            const VecD &gen_coeffs,
                                            const VecD &parameters,
-                                           std::optional<int> only_rotate_len_k) -> void;
+                                           std::optional<size_t> only_rotate_len_k) -> void;
 
     template <typename EvolutionFunc>
     auto run_gate_loop_(const std::vector<VecZ> &majoranas,
-                        std::optional<int> only_rotate_len_k,
+                        std::optional<size_t> only_rotate_len_k,
                         EvolutionFunc evolution_func) -> void;
 
     auto propagate_one_(const VecZ &gen_vec,
-                        std::optional<int> only_rotate_len_k,
+                        std::optional<size_t> only_rotate_len_k,
                         std::optional<std::reference_wrapper<const VecD>> coeffs = std::nullopt,
                         std::optional<double> param = std::nullopt,
                         size_t param_index = 0,
@@ -449,7 +449,7 @@ private:
     // fused_scale_coeffs (ContractImmediately only): the picture's mutable coeff vector for the uncapped
     // fused cos sweep; the taken decision is reported via fused_scale so the apply matches. See build_layer.
     auto build_evolve_result_(const VecZ &gen_vec,
-                              std::optional<int> only_rotate_len_k,
+                              std::optional<size_t> only_rotate_len_k,
                               std::optional<std::reference_wrapper<const VecD>> coeffs = std::nullopt,
                               std::optional<double> param = std::nullopt,
                               CosMask *out_cos = nullptr,
